requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
mediapipe==0.10.32
pillow==12.0.0
numpy==2.3.5
pandas==3.0.1
matplotlib==3.10.8
scikit-learn==1.8.0
seaborn==0.13.2
python-multipart==0.0.22
opencv-python==4.13.0.92